Non-uniform mesh
This version is capable of handling a non-uniform mesh. The mesh is generated by combining and sorting a uniform mesh and the grid mesh locations. All calculations are based on the current MeshStep, and therefore, does not depend on a uniform mesh like previous versions did. 
power.py was added with function Initial to provide an array of the initial power in each cell. This was to adapt thermal.py to run with an external power input.