Power tally finished
Power tally is completed in Power and used in Channel. Produces an arbitrarily large number, scaled down to conserve expected fuel pin power by a factor of 0.00015. In Channel, power is divided by MeshStep (height of the cell) to produce a linear power. 
Currently trying to create the cell & edge consistency. Right now, the Tally has a 0 inserted at the bottom to make it a long as the mesh edge vectors, but Channel needs to be modified to use cells values.